requirements.txt
Django==1.6
Pillow==2.5.3
South==1.0
djangorestframework==2.4.3
